dh_clideps: Add dependencies required for .installcliframework handling.
This adds a dependency on cli-common >= 0.9 (the first release with framework support)
and checks that the assemblies referenced in the .installcliframework file are provided
by a binary package the source generates. If the framework file references an assembly
from a different binary package, add a strictly versioned dependency on that package.